Us5 3

+++ src/controllers/cartController.js
@@ -3,6 +3,11 @@ const { contadorCarrito } = require('../models/carrtio');
 const { stockDown } = require('../models/products');
 
 function addToCart(req, res) {
+    const productId = productsService.normalizeId(req.body.productId);
+    if(productId === null){
+        return res.status(400).send('ID no valido');
+    }
+    const producto = productsService.getProductById(productId);
     const productId = Number(req.body.productId);
     const resultado = cartService.addProduct(req.session, productId);
 
@@ -29,19 +34,47 @@ function cargarCarrito(req, res) {
 }
 
 function incrementarCantidad(req, res) {
-    const productId = Number(req.params.id);
+    const productId = productsService.normalizeId(req.params.id);
+    if (productId === null) {
+        return res.status(400).send('ID no valido');
+    }
+
+    const producto = productsService.getProductById(productId);
+
+    if(!producto) {
+        return res.status(404).send('Producto no encontrado');
+    }
+
     cartService.incrementQuantity(req.session, productId);
     res.redirect('/cart/list');
 }
 
 function decrementarCantidad(req, res) {
-    const productId = Number(req.params.id);
+    const productId = productsService.normalizeId(req.params.id);
+    if(productId === null) {
+        return res.status(400).send('ID no valido');
+    }
+
+    const producto = productsService.getProductById(productId);
+
+    if(!producto){ 
+        return res.status(404).send('Producto no encontrado');
+    }
     cartService.decrementQuantity(req.session, productId);
     res.redirect('/cart/list');
 }
 
 function quitarProducto(req, res) {
-    const productId = Number(req.params.id);
+    const productId = productsService.normalizeId(req.params.id);
+    if(productId === null) {
+        return res.status(400).send('ID no valido');
+    }
+
+    const producto = productsService.getProductById(productId);
+
+    if(!producto){ 
+        return res.status(404).send('Producto no encontrado');
+    }
     cartService.removeProduct(req.session, productId);
     res.redirect('/cart/list');
 }
